# docs/swagger.yaml
definitions:
  main.Task:
    properties:
      done:
        type: boolean
      id:
        type: integer
      title:
        type: string
    type: object
info:
  contact: {}
paths:
  /ping:
    get:
      description: Responds with pong to show the server is running
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - health
  /tasks:
    get:
      description: Returns a list of all tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Task'
            type: array
      summary: Get all tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Adds a new task to the list
      parameters:
      - description: Task to add
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/main.Task'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add a new task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Deletes a task by its ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete task by ID
      tags:
      - tasks
    get:
      description: Returns a single task by its ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Task'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get task by ID
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Updates a task by its ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/main.Task'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update task by ID
      tags:
      - tasks
swagger: "2.0"